@@ -1,10 +1,11 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import TodoList from './components/TodoList/TodoList';
 import './components/TodoList/TodoList.css';
 import * as serviceWorker from './serviceWorker';
 
-ReactDOM.render(<TodoList />, document.getElementById('root'));
+const root = createRoot(document.getElementById('root'));
+root.render(<TodoList />);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
